requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
Django==2.2.24
django-crispy-forms==1.11.2
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.0.4
isort==5.8.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow>=8.2.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptvsd==4.3.2
pytz==2020.5
PyYAML==5.4.1
regex==2021.3.17
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
